# feat(query): Query Planning optimization for 'or vector(0)'