**Pull Request checklist**

- [x] The commit(s) message(s) follows the contribution [guidelines](CONTRIBUTING.md) ?
- [x] Tests for the changes have been added (for bug fixes / features) ?
- [ ] Docs have been added / updated (for bug fixes / features) ?

`foo or vector(0)` like queries are very common. Current plan materialization results in SetOperatorExec which is expensive since it transmits data over the network and uses more memory. 

A more optimal implementation would be to apply an InstantFunctionMapper to convert the NaNs to 0s. This plan is memory efficient and does not write results to the wire unnecessarily. This is what this PR delivers.

Unit tests cover query planning.

Local performance test shows 25% reduction in average latency. 

Gatling Test Results were run for query `heap_usage0{_ns_="App-0",_ws_="demo", host="H1"} or vector(0)`

1 FiloDB node, 8 shards
2G heap, M1 mac
10 queries per second
Cardinality of result 1250
Time range queried: 3h
Step: 60s


## With optimization

### Run 1
```
================================================================================
---- Global Information --------------------------------------------------------
> request count                                       1800 (OK=1800   KO=0     )
> min response time                                    126 (OK=126    KO=-     )
> max response time                                    358 (OK=358    KO=-     )
> mean response time                                   135 (OK=135    KO=-     )
> std deviation                                         12 (OK=12     KO=-     )
> response time 50th percentile                        131 (OK=131    KO=-     )
> response time 75th percentile                        135 (OK=135    KO=-     )
> response time 95th percentile                        149 (OK=149    KO=-     )
> response time 99th percentile                        160 (OK=160    KO=-     )
> mean requests/sec                                  9.945 (OK=9.945  KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          1800 (100%)
> 800 ms < t < 1200 ms                                   0 (  0%)
> t > 1200 ms                                            0 (  0%)
> failed                                                 0 (  0%)
================================================================================
```
### Run 2
```
================================================================================
---- Global Information --------------------------------------------------------
> request count                                       1800 (OK=1800   KO=0     )
> min response time                                    125 (OK=125    KO=-     )
> max response time                                    435 (OK=435    KO=-     )
> mean response time                                   134 (OK=134    KO=-     )
> std deviation                                         11 (OK=11     KO=-     )
> response time 50th percentile                        130 (OK=130    KO=-     )
> response time 75th percentile                        134 (OK=134    KO=-     )
> response time 95th percentile                        149 (OK=149    KO=-     )
> response time 99th percentile                        165 (OK=165    KO=-     )
> mean requests/sec                                  9.945 (OK=9.945  KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          1800 (100%)
> 800 ms < t < 1200 ms                                   0 (  0%)
> t > 1200 ms                                            0 (  0%)
> failed                                                 0 (  0%)
================================================================================
```
### Run 3

```
================================================================================
---- Global Information --------------------------------------------------------
> request count                                       1800 (OK=1800   KO=0     )
> min response time                                    125 (OK=125    KO=-     )
> max response time                                    334 (OK=334    KO=-     )
> mean response time                                   134 (OK=134    KO=-     )
> std deviation                                         11 (OK=11     KO=-     )
> response time 50th percentile                        130 (OK=130    KO=-     )
> response time 75th percentile                        134 (OK=134    KO=-     )
> response time 95th percentile                        149 (OK=149    KO=-     )
> response time 99th percentile                        168 (OK=168    KO=-     )
> mean requests/sec                                  9.945 (OK=9.945  KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          1800 (100%)
> 800 ms < t < 1200 ms                                   0 (  0%)
> t > 1200 ms                                            0 (  0%)
> failed                                                 0 (  0%)
================================================================================
```

## Without optimization

### Run 

```
================================================================================
---- Global Information --------------------------------------------------------
> request count                                       1800 (OK=1800   KO=0     )
> min response time                                      6 (OK=6      KO=-     )
> max response time                                    920 (OK=920    KO=-     )
> mean response time                                   179 (OK=179    KO=-     )
> std deviation                                         48 (OK=48     KO=-     )
> response time 50th percentile                        180 (OK=180    KO=-     )
> response time 75th percentile                        192 (OK=192    KO=-     )
> response time 95th percentile                        217 (OK=217    KO=-     )
> response time 99th percentile                        302 (OK=302    KO=-     )
> mean requests/sec                                  9.945 (OK=9.945  KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          1798 (100%)
> 800 ms < t < 1200 ms                                   2 (  0%)
> t > 1200 ms                                            0 (  0%)
> failed                                                 0 (  0%)
================================================================================
```

### Run 2

```
================================================================================
---- Global Information --------------------------------------------------------
> request count                                       1800 (OK=1800   KO=0     )
> min response time                                    133 (OK=133    KO=-     )
> max response time                                    500 (OK=500    KO=-     )
> mean response time                                   180 (OK=180    KO=-     )
> std deviation                                         29 (OK=29     KO=-     )
> response time 50th percentile                        178 (OK=178    KO=-     )
> response time 75th percentile                        190 (OK=190    KO=-     )
> response time 95th percentile                        210 (OK=210    KO=-     )
> response time 99th percentile                        283 (OK=283    KO=-     )
> mean requests/sec                                  9.945 (OK=9.945  KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          1800 (100%)
> 800 ms < t < 1200 ms                                   0 (  0%)
> t > 1200 ms                                            0 (  0%)
> failed                                                 0 (  0%)
================================================================================
```

### Run 3

```
================================================================================
---- Global Information --------------------------------------------------------
> request count                                       1800 (OK=1800   KO=0     )
> min response time                                    134 (OK=134    KO=-     )
> max response time                                    455 (OK=455    KO=-     )
> mean response time                                   181 (OK=181    KO=-     )
> std deviation                                         21 (OK=21     KO=-     )
> response time 50th percentile                        180 (OK=180    KO=-     )
> response time 75th percentile                        192 (OK=192    KO=-     )
> response time 95th percentile                        213 (OK=212    KO=-     )
> response time 99th percentile                        233 (OK=233    KO=-     )
> mean requests/sec                                  9.945 (OK=9.945  KO=-     )
---- Response Time Distribution ------------------------------------------------
> t < 800 ms                                          1800 (100%)
> 800 ms < t < 1200 ms                                   0 (  0%)
> t > 1200 ms                                            0 (  0%)
> failed                                                 0 (  0%)
================================================================================
```